Everything but edit item working.

+++ week3/goldapp/js/main.js
@@ -1,7 +1,7 @@
 var parseLogItemForm = function(data){
 	// uses form data here;
 	console.log(data);
-}
+};
 
 
 
@@ -41,58 +41,39 @@ $(document).bind('pageinit', function(){
 			//Converting string from local storage value back to an object using JSON.parse()
 			var obj = JSON.parse(value);
 			//create log item list
-			var optSubText = $( "img src='images/"+logItem.treatments[1]+".jpg'/>"+
-				"<h3>"+logItem.date[1]+"</h3>"+
-				"<h3>"+logItem.currentTime[1]+"</h3>"+
-				"<p>"+logItem.fname[0]+": "+logItem.fname[1]+"</p>"+
-				"<p>"+logItem.lname[0]+": "+logItem.lname[1]+"</p>"+
-				"<p>"+logItem.bsreading[0]+": "+logItem.bsreading[1]+"</p>"+
-				"<p>"+logItem.sex[0]+": "+logItem.sex[1]+"</p>"+
-				"<p>"+logItem.condition[0]+": "+logItem.condition[1]+"</p>"+
-				"<p>"+logItem.treatments[0]+": "+logItem.treatments[1]+"</p>"+
-				"<p>"+logItem.comments[0]+": "+logItem.comments[1]+"</p>");
+			var optSubText = $( "<img src='images/"+obj.treatments[1]+".jpg'/>"+
+				"<h3>"+obj.date[1]+"</h3>"+
+				"<h3>"+obj.currentTime[1]+"</h3>"+
+				"<p>"+obj.fname[0]+" "+obj.fname[1]+"</p>"+
+				"<p>"+obj.lname[0]+" "+obj.lname[1]+"</p>"+
+				"<p>"+obj.bsreading[0]+" "+obj.bsreading[1]+"</p>"+
+				"<p>"+obj.sex[0]+" "+obj.sex[1]+"</p>"+
+				"<p>"+obj.condition[0]+" "+obj.condition[1]+"</p>"+
+				"<p>"+obj.treatments[0]+" "+obj.treatments[1]+"</p>"+
+				"<p>"+obj.comments[0]+" "+obj.comments[1]+"</p>");
+			//Creating Edit Link in Item
+			var editLink = $("<a href='#add' id='edit"+key+"'> Edit Log Item</a>");
+				editLink.bind('click', function(){
+					editItem(this.id);
+				});
+			//Creating Delete Link in Item
+			var deleteLink = $("<a href='#list' id='delete"+key+"'>Delete Item</a>");
+				deleteLink.bind('click', function(){
+					deleteItem(this.id);
+				});
+			//Make item data the edit link
+			editLink.html(optSubText);
+			//Adding edit and delete links to the list
+			makeli.append(editLink, deleteLink).appendTo("#logitemList");
 			}
-			makeItemLinks(localStorage.key(i), linksLi);//Create edit and delete links for each item in local storage.
-		}
-	}
-});
-/*
-	//Function to create the edit and delete item links for each item in local storage.
-	function makeItemLinks(key, linksLi) {
-		//add edit single item link
-		var editLink = document.createElement("a");
-		editLink.href = "#";
-		editLink.key = key;
-		var editText = "Edit Log Entry";
-		editLink.addEventListener("click", editItem);
-		editLink.innerHTML = editText;
-		linksLi.appendChild(editLink);
+	$("#logitemList").listview('refresh');			
+	};
 
-		
-		// add line break
-		var breakTag = document.createElement("br");
-		linksLi.appendChild(breakTag);
-
-		//add delete single item link
-		var deleteLink = document.createElement("a");
-		deleteLink.href = "#";
-		deleteLink.key = key;
-		var deleteText = "Delete Log Entry";
-		deleteLink.addEventListener("click", deleteItem);
-		deleteLink.innerHTML = deleteText;
-		linksLi.appendChild(deleteLink);
-
-
-	}
 	//edit single item
-	function editItem() {
+	var editItem =function() {
 		//grab the data from our item in local storage
 		var value = localStorage.getItem(this.key);
 		var logItem = JSON.parse(value);
-
-		//show form
-		toggleControls("off");
-
 		//Populate the form with current local storage values.
 		ge("fname").value = logItem.fname[1];
 		ge("lname").value = logItem.lname[1];
@@ -110,35 +91,66 @@ $(document).bind('pageinit', function(){
 		ge("condition").value = logItem.condition[1];
 		ge("treatments").value = logItem.treatments[1];
 		ge("comments").value = logItem.comments[1];
-
-		//remove initial listener from the input "save log item" button
-		submitLink.removeEventListener("click", storeData);
 		//Change submit button value to edit button
-		ge("submit").value = "Edit Log Entry";
-		var editSubmit = ge("submit");
-		//Save the key value established in this vunction as a property of the editSubmit event
-		//so we can use that value when we save the data
-		editSubmit.addEventListener("click", validate);
-		editSubmit.key = this.key;
-	}
-	function deleteItem(){
+		$("span.ui-controlgroup-last").html("Edit Log Item");
+		//Save the key value established in this vunction as a property of #addLogItem
+		$("#addLogItem").attr("key", key);
+		//Refresh the menu
+		$("select#treatment").selectmenu("refresh");
+	};
+	
+	var deleteItem = function(){
 		var ask = confirm("Are you sure you want to delete this log entry?");
+		var key = localStorage.key();
 		if(ask){
-			localStorage.removeItem(this.key);
-			window.location.reload();
+			localStorage.removeItem(key);
 			alert("Log Entry was deleted.");
+			getData();
 		}else{
 			alert("Log entry was Not deleted.");
-		}
-	}
+		};
+	};
+
+	//Auto Populate Default data to local storage
+	var autoFillData = function(){
+		//Store the JSON Object into local storage
+		for(var n in json){
+			var id = Math.floor(Math.random()*100000001);
+			localStorage.setItem(id, JSON.stringify(json[n]));
+		};
+	
+	};
+
 	//clear local storage
-	function clearData() {
+	var clearData = function() {
 		if(localStorage.length === 0){
 			alert("There is no data to clear.");
 		}else{
-			localStorage.clear();
-			alert("All log items are deleted!");
-			window.location.reload();
-			return false;
-		}
-	}*/
+			var ask = confirm("Deleting ALL log items? This can NOT be undone.");
+			if(ask){
+				localStorage.clear();
+				alert("All log items are deleted!");
+				return false;
+			}else{
+				alert("Log items not deleted.");
+			};
+		};
+	};
+
+
+	//Set Link and Submint Click Events
+	var displayLink = ge("displayLog");
+	displayLink.addEventListener("click", getData);
+
+	var clearLink = ge("clear");
+	clearLink.addEventListener("click", clearData);
+	
+/*	var submitLink = ge("submit");
+	submitLink.addEventListener("click", validate);*/
+
+});
+/*
+	
+	
+	
+	*/
